Updated because of Task API changes. Changed Callable<Integer> to Callable<Long> everywhere. Started work on Task Manager. Added Callable debug action to debug dialog.
git-svn-id: https://lilith.svn.sourceforge.net/svnroot/lilith/trunk@163 4afeb0b7-f550-0410-824e-ec7672375c7e

+++ lilith/src/main/java/de/huxhorn/lilith/Lilith.java
@@ -322,7 +322,7 @@ public class Lilith
 				callable.addPropertyChangeListener(new IndexingChangeListener());
 				try
 				{
-					int count = callable.call();
+					long count = callable.call();
 					if(logger.isInfoEnabled())
 					{
 						logger.info("Finished indexing {}. Number of events: {}", logFile.getAbsolutePath(), count);
@@ -594,7 +594,7 @@ public class Lilith
 
 		public CreateMainFrameRunnable(ApplicationPreferences applicationPreferences, SplashScreen splashScreen, String appTitle, boolean enableBonjour)
 		{
-			this.splashScreen=splashScreen;
+			this.splashScreen = splashScreen;
 			this.enableBonjour = enableBonjour;
 			this.appTitle = appTitle;
 			this.applicationPreferences = applicationPreferences;
